pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safety_polytope"
version = "0.1.0"
description = "Learning Safety Constraints for Large Language Models (Safety Polytope)"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.10",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "accelerate==0.30.0",
  "anthropic",
  "art",
  "boto3",
  "bpe",
  "confection==0.1.4",
  "datasets==2.20.0",
  "datasketch",
  "fschat",
  "google-cloud-aiplatform",
  "google-generativeai",
  "hydra-core==1.3.2",
  "hydra-submitit-launcher==1.2.0",
  "kernels==0.10.2",
  "loguru>=0.7.3",
  "matplotlib",
  "mistralai",
  "nltk==3.9.1",
  "omegaconf==2.3.0",
  "openai>=1.25.1",
  "pandas==2.2.3",
  "pip",
  "ray",
  "scikit-learn==1.6.1",
  "scipy==1.15.2",
  "seaborn==0.13.2",
  "sentence_transformers",
  "spacy>=3.7.2",
  "streamlit>=1.50.0",
  "tenacity",
  "tensorflow==2.18.0",
  "torch",
  "torchvision",
  "tqdm==4.66.4",
  "transformers>=4.45.2",
  "triton==3.4.0",
  "tueplots==0.0.17",
  "umap-learn==0.5.6",
  "vllm>=0.6.3.post1",
]

[project.optional-dependencies]
dev = [
  "black>=23.1.0",
  "isort>=5.12.0",
  "flake8>=6.0.0",
  "pytest>=7.0.0",
  "pre-commit>=3.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
zip-safe = false

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.extra-build-dependencies]
# Ensure xformers builds with torch available in build isolation
xformers = ["torch"]
